requirements.txt
python-docx>=1.1.0
marker-pdf>=1.10.2